==> code/prep/rosmap/11_download_rosmap_clinical_synapse.R <==
library(synapser)
library(synapserutils)

synapser::synLogin()

# download rosmap clinical data set
syncFromSynapse('syn3191087', 
                path="data/synapse", 
                ifcollision="keep.local")
